requirements.txt
fastapi
uvicorn
langchain
langchain_openai
langchain_community
langgraph
httpx
python-dotenv
numpy>=1.21.0
openai>=1.0.0